Point tests to relocated functions
